Fix the storage of the previous weather lookup when used with id: prefixes
The handling of id: lookups rewrites the location variable prior to its storage,
meaning that the "id:" prefix has been stripped away and when the value is
recalled, the lookup is treated as a zip code instead. This commit resolves
this issue by keeping a separate unaltered variable for the value to be used
with storage.
This functionality will also benefit any other prefix-style lookups in the
future, such as a lat/lon lookup prefix.